Faxable CCPA-only letter
Letter plus Makefile changes to support it.

To use:

 * Set the CCPA environment variables

 * `make ccpa-fax.pdf` and print it

 * Fax it.

Note: if the company does not have their fax number on their "contact
us" page, check Investor Relations, the annual report, or documents
they're required to file with regulators.

Don't forget to plug the regular phone back in after using the fax
machine.  (Or leave the fax machine plugged in at peak telemarketing
hours to see if they put it down as a fax line and stop calling.)