Add an "any" activity column
Add a separate column for count of total users whose info the company
has sent to Facebook in any way.

Before, it was hard to tell if (a) a company had 2 users in custom audiences
and a different 2 users in CAPI, or (b) both custom audiences and CAPI for
the same 2 users.

Now, the (a) example will have a count of 4 and the (b) example will
have a count of 2.

Should help to measure total presence and/or impact of a company.